+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This file documents notable changes to the project
 **new**
 
 - Set a response body when creating a `HTTPError`, either via the constructor or the static methods.
+- Structure primatives' fallback is now optional. If a fallback isn't provided, validation will fail if with a "Missing value" if no value is provided.
+- Added an unstable/experimental `Structure.array` for validating an array of a single Structure, e.g. an array of strings.
+- Added `config.number(...)` & `config.boolean(...)` types along with `Structure` equivolents.
 
 ## 0.3.0